check the makefile and correct whatever errors it gievs

+++ ft_memchr.c
@@ -1,6 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memchr.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ielabdal <ielabdal@learner.42.tech>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/09/19 18:26:20 by ielabdal          #+#    #+#             */
+/*   Updated: 2026/09/20 17:26:37 by ielabdal         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
-#include <stddef.h>
-void *ft_memchr(const void *ptr,int c,unsigned int n)
+void *ft_memchr(const void *ptr,int c,size_t n)
 {
 	unsigned char ch;
 	const unsigned char *str;
