+++ ft_strlcpy.c
@@ -1,5 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ielabdal <ielabdal@learner.42.tech>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/09/19 18:38:48 by ielabdal          #+#    #+#             */
+/*   Updated: 2026/09/20 16:50:19 by ielabdal         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
-unsigned int	ft_strlcpy(char *dest,const char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest,const char *src, size_t size)
 {
 	int	len;
